src/product/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
prometheus-fastapi-instrumentator
structlog
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
